Works on the Teensy

+++ src/dbgcfg.cpp
@@ -19,16 +19,17 @@ void dumpHex(uint32_t v, const char* header) {
 }
 
 void dumpHex(uint16_t v, const char* header) {
-  dumpHex(static_cast<unsigned long>(v), header);
+  dumpHex(static_cast<uint32_t>(v), header);
 }
 
 void dumpHex(uint8_t v, const char* header) {
-  dumpHex(static_cast<unsigned long>(v), header);
+  dumpHex(static_cast<uint32_t>(v), header);
 }
 
 void dumpHex(bool v, const char* header) {
-  dumpHex(static_cast<unsigned long>(v), header);
+  dumpHex(static_cast<uint32_t>(v), header);
 }
+
 void dumpStr(const char *str) {
   Serial.print(str);
 }
@@ -37,10 +38,10 @@ const char* zeros = "00000000";
 void dumpHex(uint64_t v, const char* header) {
   if (header)
     Serial.print(header);
-  Serial.print(static_cast<unsigned long>(v >> 32), HEX);
+  Serial.print(static_cast<uint32_t>(v >> 32), HEX);
   Serial.print("|");
   // Zero-pad to the left
-  uint32_t lo = static_cast<unsigned long>(v);
+  uint32_t lo = static_cast<uint32_t>(v);
   Serial.print(&zeros[(flsl(lo) + 3) >> 2]);
   if (lo) {
     Serial.println(lo, HEX);
